chore: splitted the code for project creation form
