+++ web/ce/types/projects/projects.ts
@@ -1,0 +1,3 @@
+import { IProject } from "@plane/types";
+
+export type TProject = IProject;
