+++ web/core/components/project/index.ts
@@ -5,7 +5,6 @@ export * from "./settings";
 export * from "./card-list";
 export * from "./card";
 export * from "./create-project-modal";
-export * from "./create-project-form";
 export * from "./project-feature-update";
 export * from "./delete-project-modal";
 export * from "./form-loader";
@@ -19,3 +18,4 @@ export * from "./member-list-item";
 export * from "./project-settings-member-defaults";
 export * from "./send-project-invitation-modal";
 export * from "./confirm-project-member-remove";
+export * from "@/plane-web/components/projects/create/root";
