+++ web/ee/components/projects/create/attributes.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/projects/create/attributes";
