+++ web/ee/components/projects/create/root.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/projects/create/root";
