+++ web/ee/types/projects/projects.ts
@@ -1,0 +1,1 @@
+export * from "ce/types/projects/projects";
